+++ src/apis/http/index.js
@@ -3,6 +3,7 @@ import bank from './bank';
 import blocks from './blocks';
 import chart from './chart';
 import commercio from './commercio';
+import cosmwasm from './cosmwasm';
 import distribution from './distribution';
 import gaiaRest from './gaiaRest';
 import governance from './governance';
@@ -21,6 +22,7 @@ export {
   blocks,
   chart,
   commercio,
+  cosmwasm,
   distribution,
   gaiaRest,
   governance,