+++ src/store/keplr/getters.js
@@ -4,4 +4,6 @@ export default {
   isInitialized: (state) => state.isInitialized,
   hasKeplr: (state) => state.hasKeplr,
   accounts: (state) => state.accounts,
+  wallet: (state) =>
+    state.accounts.length > 0 ? state.accounts[0]['address'] : '',
 };