+++ src/utils/index.js
@@ -2,10 +2,12 @@ import bech32Manager from './bech32Manager';
 import blocksRequestHelper from './blocksRequestHelper';
 import coinAdapter from './coinAdapter';
 import dateHandler from './dateHandler';
+import msgBuilder from './msgBuilder';
 import proposerHandler from './proposerHandler';
 import regExpBuilder from './regExpBuilder';
 import storageHandler from './storageHandler';
 import stringEncoder from './stringEncoder';
+import tokensHandler from './tokensHandler';
 import txHandler from './txs/txHandler';
 import txsFilter from './txs/txsFilter';
 import txsTableAdapter from './txs/txsTableAdapter';
@@ -15,10 +17,12 @@ export {
   blocksRequestHelper,
   coinAdapter,
   dateHandler,
+  msgBuilder,
   proposerHandler,
   regExpBuilder,
   storageHandler,
   stringEncoder,
+  tokensHandler,
   txHandler,
   txsFilter,
   txsTableAdapter,